requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0
httpx>=0.25.0